functions/export_reviews/requirements.txt
google-cloud-firestore==2.16.1
